(WIP) Checking in initial cut of EFL support
Configures, doesn't compile.  The current codebase commits
a sin of building INTRIM code generators that try to  run
on the Host but when compiled for alternate targets blows
up in your face.

Need to see if the old stuff corrects for this BS (yes,
it's **BS**) or not...  Annoying.  Was hoping to have
efl cross-compiling so we can get E-22 up and runing on
cross-compile targets.