Oh, yeaaaah!   efl compiles and packages!
Still a WIP, but the thing now passes the next stated
milestone.  We got it sorted out for build with everything
mostly useful turned on except bullet.  We'll eventually
revisit that as a feature turn on/off item as well as
a few others.  But now, on to specifying the right
packaging.