Correct configuration for EFL as nativesdk
Well, as best as I can tell for right now.  Might need to involve
the community over there to help make sure we can make this as
clean and efficient as we possibly can down the line- but for
now, it looks like it grabs the right utils for cross compile.