Move to EFL/Enlighenment latest
- This is not the full metadata move.  This is to move the critical
  things forward into the future.  Namely EFL, Enlightenment, the
  Python bindings relative to ourselves, and a few adversely impacted
  apps.  This does NOT lead to the apps working right yet.  This is
  to blaze a trail to Gatesgarth for myself and and a few others.
- EFL moved to 1.25.1
- Python 3 EFL moved to 1.25.0
- Enlightenment moved to 0.22.4
- Entrance moved to tip of repo.
- Removed the Python extras we needed to get this all to work
  (XDG, dbus...)
- Removed all Python 2.7 support as this is sunsetted

As the last item kind of forces you to Dunfell and later, the layer
no longer supports anything prior.  As getting to Gatesgarth unlocks
quite a few important things (REAL support for a Xavier NX out of
meta-tegra, which includes CUDA 10.1 support proper...) we're caring
a bit less than usual for this one.  We want the close to the latest
EFL/Enlightenment as we can get and we can slowly fix/patch to the
rest once we don't let the magic smoke out of a machine with this
set of semi-major changes.